升级boot2.0，加入redisson并发控制

+++ src/main/java/com/files/controller/FileClientController.java
@@ -2,12 +2,17 @@ package com.files.controller;
 
 import com.files.client.FileInfo;
 import com.files.client.Result;
+import com.files.enums.ResponseEnum;
 import com.files.service.FileService;
 import lombok.extern.slf4j.Slf4j;
+import org.redisson.Redisson;
+import org.redisson.api.RLock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.File;
+import java.util.concurrent.TimeUnit;
+
 /**
  * 微服务对应接口
  * @author fhx
@@ -18,7 +23,9 @@ import java.io.File;
 public class FileClientController {
 	@Autowired
 	private FileService tmpFileService;
-	
+	@Autowired
+	private Redisson redisson;
+
 	/**
 	 * 文件删除
 	 * @param fileInfo
@@ -27,13 +34,26 @@ public class FileClientController {
 	@RequestMapping(value = "/file/delete", method = RequestMethod.POST)
 	public Result<?> deleteFile(@RequestBody FileInfo fileInfo) {
 		Result<Long> result = Result.create(true);
+		RLock lock = redisson.getLock("seckill:" + fileInfo.getFileName());
 		try {
-			tmpFileService.deleteFile(fileInfo);
-		} catch (Exception e) {
-			log.error("failed to delete file:{}", fileInfo, e);
+			// 并发请求
+			if (lock.tryLock(2, 10, TimeUnit.SECONDS)) {
+				tmpFileService.deleteFile(fileInfo);
+			}
+		} catch (InterruptedException e) {
+			log.error("获取分布式锁异常", e);
 			result.setCode("999999");
 			result.setMessages("文件删除异常！");
+		}  catch (Exception e) {
+			log.error("failed to delete file:{}", fileInfo, e);
+			result.setCode(ResponseEnum.SYS_FAILD.getCode());
+			result.setMessages("文件删除异常！");
+		} finally {
+			if (lock.isHeldByCurrentThread()) {
+				lock.unlock();
+			}
 		}
+
 		return result;
 	}
 	
@@ -46,14 +66,28 @@ public class FileClientController {
 	@RequestMapping(value = "/file/switchgroup", method = RequestMethod.POST)
 	public Result<?> switchFileGroup(@RequestBody FileInfo tmpFileInfo, @RequestParam String targetGroup) {
 		log.info("======文件组移动：{}--->{}=========",tmpFileInfo.getGroup(), targetGroup);
-		Result<?> result = Result.create(true);;
+		Result<?> result = Result.create(true);
+
+		RLock lock = redisson.getLock("seckill:" + tmpFileInfo.getFileName());
 		try {
-			tmpFileService.switchFileGroup(tmpFileInfo, targetGroup);
-		} catch (Exception e) {
-			log.error("failed to switch file:{} to group:{}", tmpFileInfo, targetGroup, e);
+			// 并发请求
+			if (lock.tryLock(2, 10, TimeUnit.SECONDS)) {
+				tmpFileService.switchFileGroup(tmpFileInfo, targetGroup);
+			}
+		} catch (InterruptedException e) {
+			log.error("获取分布式锁异常", e);
 			result.setCode("999999");
+			result.setMessages("文件删除异常！");
+		}  catch (Exception e) {
+			log.error("failed to switch file:{} to group:{}", tmpFileInfo, targetGroup, e);
+			result.setCode(ResponseEnum.SYS_FAILD.getCode());
 			result.setMessages("文件移动异常！");
+		} finally {
+			if (lock.isHeldByCurrentThread()) {
+				lock.unlock();
+			}
 		}
+
 		return result;
 	}
